+++ src/kentaasvang.Rssh/Implementations/Commands/ListCommandWrapper.cs
@@ -1,8 +1,9 @@
 using System;
 using System.CommandLine;
-using kentaasvang.Rssh.Interfaces;
+using kentaasvang.Rssh.Interfaces.Commands;
+using kentaasvang.Rssh.Interfaces.Handlers;
 
-namespace kentaasvang.Rssh.Commands;
+namespace kentaasvang.Rssh.Implementations.Commands;
 
 public class ListCommandWrapper : ICommandWrapper
 {
