+++ src/kentaasvang.Rssh/Implementations/Commands/RemoveCommandWrapper.cs
@@ -1,8 +1,8 @@
 using System.CommandLine;
-using kentaasvang.Rssh.Data;
-using kentaasvang.Rssh.Interfaces;
+using kentaasvang.Rssh.Interfaces.Commands;
+using kentaasvang.Rssh.Interfaces.Handlers;
 
-namespace kentaasvang.Rssh.Commands;
+namespace kentaasvang.Rssh.Implementations.Commands;
 
 public class RemoveCommandWrapper : ICommandWrapper
 {
