+++ src/kentaasvang.Rssh/Interfaces/Handlers/IListHandler.cs
@@ -1,4 +1,4 @@
-namespace kentaasvang.Rssh.Interfaces;
+namespace kentaasvang.Rssh.Interfaces.Handlers;
 
 public interface IListHandler
 {
