--- src/kentaasvang.Rssh/Interfaces/ICommandInstaller.cs
@@ -1,8 +1,0 @@
-using System.CommandLine;
-
-namespace kentaasvang.Rssh.Interfaces;
-
-internal interface ICommandInstaller
-{
-    Command LoadCommand();
-}
