+++ tests/kentaasvang.Rssh.UnitTests/AddHandlerTest.cs
@@ -1,9 +1,10 @@
 namespace kentaasvang.Rssh.UnitTests;
 
-using kentaasvang.Rssh.Implementations;
 using kentaasvang.Rssh.Interfaces;
 using kentaasvang.Rssh.Entities;
 using kentaasvang.Rssh.Repositories;
+using kentaasvang.Rssh.Interfaces.Handlers;
+using kentaasvang.Rssh.Implementations.Handlers;
 
 public class AddHandlerTest
 {
